requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
spacy==3.5.0
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0.tar.gz
scikit-learn==1.2.2
numpy==1.23.5
google-ads>=28.0.0,<29
pandas==2.3.3
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.2.1
google-api-python-client==2.161.0
oauth2client==4.1.3
openai==1.68.2
gunicorn==21.2.0
firebase-admin==6.7.0

